+++ frictionless/plugins/pandas/__init__.py
@@ -1,0 +1,3 @@
+from .plugin import PandasPlugin
+from .dialect import PandasDialect
+from .parser import PandasParser
